requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
google-generativeai>=0.3.0
httpx>=0.23.0
firebase-admin>=6.0.0
python-dotenv>=0.19.0
python-multipart>=0.0.6
groq>=0.3.0  